added Flask blueprints for the ext-recuit endpoint
