+++ user_interface/app.py
@@ -1,16 +1,22 @@
+# app.py
 import os
 import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-import sqlite3
-from flask import Flask, flash, redirect, render_template, request, session, jsonify, send_from_directory, g, url_for
-from flask_session import Session
-from werkzeug.security import check_password_hash, generate_password_hash
-from werkzeug.utils import secure_filename
-from helpers import apology_login, login_required
-import main
+# Add the parent directory to sys.path
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
 import pandas as pd
 import json
-import datetime
+import sqlite3
+from flask import Flask, flash, redirect, render_template, request, session, send_from_directory, g, url_for, current_app
+from flask_session import Session
+from werkzeug.security import check_password_hash, generate_password_hash
+
+from helpers import apology_login, login_required, get_db
+import main
+from main import get_resumes, resume_extraction, job_info_extraction, calc_similarity
+
+# Import the blueprint from the recruitment package
+from recruitment import recruitment_bp
 
 # Configure application
 app = Flask(__name__)
@@ -20,38 +26,25 @@ app.config["SESSION_PERMANENT"] = False
 app.config["SESSION_TYPE"] = "filesystem"
 Session(app)
 
-# Database configuration
-DATABASE = 'skills.db'
+# File upload configuration
+UPLOAD_FOLDER = os.path.join(os.getcwd(), 'uploads')
+if not os.path.exists(UPLOAD_FOLDER):
+    os.makedirs(UPLOAD_FOLDER)
+app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
-def get_db():
-    db = getattr(g, '_database', None)
-    if db is None:
-        db = g._database = sqlite3.connect(DATABASE)
-        db.execute('PRAGMA foreign_keys = ON;')
-    return db
+# Register the blueprint
+app.register_blueprint(recruitment_bp)
 
+# Database teardown function
 @app.teardown_appcontext
-def close_connection(exception):
-    db = getattr(g, '_database', None)
+def close_db(exception):
+    db = g.pop('db', None)
     if db is not None:
         db.close()
 
-# File upload configuration
-UPLOAD_FOLDER = os.path.join(os.getcwd(), 'uploads')
-ALLOWED_EXTENSIONS = {'txt', 'pdf', 'png', 'jpg', 'jpeg', 'gif'}
-app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
-if not os.path.exists(app.config['UPLOAD_FOLDER']):
-    os.makedirs(app.config['UPLOAD_FOLDER'])
-app.config['UPLOAD_FOLDER'] = os.path.join(os.getcwd(), 'uploads')
-
-def allowed_file(filename):
-    return '.' in filename and \
-           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
-
-
 @app.after_request
 def after_request(response):
-    """Ensure responses aren't cached"""
+    """Ensure responses aren't cached."""
     response.headers["Cache-Control"] = "no-cache, no-store, must-revalidate"
     response.headers["Expires"] = 0
     response.headers["Pragma"] = "no-cache"
@@ -60,12 +53,12 @@ def after_request(response):
 @app.route("/")
 @login_required
 def index():
-    """Render Welcome Page"""
+    """Render Welcome Page."""
     return render_template("index.html")
 
 @app.route("/login", methods=["GET", "POST"])
 def login():
-    """Log user in"""
+    """Log user in."""
     session.clear()
 
     if request.method == "POST":
@@ -89,13 +82,13 @@ def login():
 
 @app.route("/logout")
 def logout():
-    """Log user out"""
+    """Log user out."""
     session.clear()
     return redirect("/")
 
 @app.route("/register", methods=["GET", "POST"])
 def register():
-    """Register user"""
+    """Register user."""
     if request.method == "POST":
         if not request.form.get("username"):
             return apology_login("must provide username", 400)
@@ -117,7 +110,7 @@ def register():
         if len(password) < 8:
             return apology_login("password must be at least 8 characters", 403)
         if len(password) > 20:
-            return apology_login("password must be less than 20 character", 403)
+            return apology_login("password must be less than 20 characters", 403)
         if not any(char.isdigit() for char in password):
             return apology_login("password must contain at least one digit", 403)
         if not any(char.isupper() for char in password):
@@ -134,191 +127,25 @@ def register():
         return redirect("/login")
     else:
         return render_template("register.html")
-    
-
-@app.route("/ext_recruit", methods=["GET", "POST"])
-@login_required
-def ext_recruit():
-    if request.method == 'POST':
-        # Determine which form was submitted
-        if 'files' in request.files:
-            # Handle file uploads
-            files = request.files.getlist('files')
-            if not files or files[0].filename == '':
-                flash('No files selected')
-                return redirect(request.url)
-
-            for file in files:
-                if allowed_file(file.filename):
-                    filename = secure_filename(file.filename)
-                    upload_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-                    file.save(upload_path)
-                else:
-                    flash(f'File type not allowed: {file.filename}')
-                    return redirect(request.url)
-
-            flash('Files successfully uploaded')
-            return redirect(url_for('ext_recruit'))
-
-        elif 'job_description' in request.form:
-            # Handle job description submission
-            job_description = request.form['job_description']
-            session['job_description'] = job_description
-            flash('Job description saved')
-            return redirect(url_for('ext_recruit'))
-
-        elif request.form.get('action') == 'run_analysis':
-            # Handle running the analysis
-            job_description = session.get('job_description')
-            upload_folder = app.config['UPLOAD_FOLDER']
-            applicant_files = [file for file in os.listdir(upload_folder) if file.endswith('.pdf')]
-
-            if not applicant_files:
-                flash('No applicants to analyze')
-                return redirect(url_for('ext_recruit'))
-
-            if not job_description:
-                flash('No job description provided')
-                return redirect(url_for('ext_recruit'))
-
-            db = get_db()
-            cursor = db.cursor()
-
-            # Step 1: Insert Job Posting into the Database
-            # Extract skills from the job description
-            df_jobs = pd.DataFrame([{'raw': job_description}])
-            df_jobs = main.job_info_extraction(df_jobs)
-
-            required_skills = df_jobs.loc[0, 'Skills']
-            required_skills_json = json.dumps(required_skills)
-
-            # Insert job posting into the database
-            cursor.execute("""
-                INSERT INTO job_postings (job_description, required_skills)
-                VALUES (?, ?)
-            """, (job_description, required_skills_json))
-            job_id = cursor.lastrowid  # Get the job_id of the inserted job
-
-            # Step 2: Insert Applicants into the Database
-            # Extract applicant data
-            df_resumes = main.get_resumes(upload_folder)
-            df_resumes = main.resume_extraction(df_resumes)
-
-            for _, row in df_resumes.iterrows():
-                name = row.get('name', '')
-                contact_info = ''  # Update as needed
-                skills = row.get('Skills', [])
-                skills_json = json.dumps(skills)
-                degrees = row.get('Degrees', [])
-                degrees_json = json.dumps(degrees)
-
-                # Insert or update applicant in the database
-                cursor.execute("""
-                    INSERT OR REPLACE INTO applicants (name, contact_info, skills, degrees)
-                    VALUES (?, ?, ?, ?)
-                """, (name, contact_info, skills_json, degrees_json))
-
-            # Commit the inserts
-            db.commit()
-
-            # Step 3: Perform Similarity Analysis
-            # Calculate similarity
-            matching_dataframe = main.calc_similarity(df_resumes, df_jobs, parallel=True)
-
-            # Update applicants with similarity scores and ranks
-            for _, row in matching_dataframe.iterrows():
-                name = row['name']
-                similarity_score = row['similarity_score']
-                rank = int(row['rank'])
-                interview_status = row['interview_status']
-
-                cursor.execute("""
-                    UPDATE applicants
-                    SET similarity_score = ?, rank = ?, interview_status = ?
-                    WHERE name = ?
-                """, (similarity_score, rank, interview_status, name))
-
-            # Commit the updates
-            db.commit()
-
-            # Step 4: Retrieve Ranked Applicants for Display
-            cursor.execute("""
-                SELECT name, similarity_score, rank, interview_status
-                FROM applicants
-                ORDER BY rank ASC
-            """)
-            ranked_applicants = cursor.fetchall()
-
-            # Prepare data for template
-            analysis_data = [{
-                'name': row[0],
-                'similarity_score': round(row[1], 3),
-                'rank': row[2],
-                'interview_status': row[3]
-            } for row in ranked_applicants]
-
-            columns = ['name', 'similarity_score', 'rank', 'interview_status']
-
-            return render_template("ext_recruit.html", applicants=applicant_files, analysis_data=analysis_data, columns=columns)
-
-    else:
-        # Handle GET request
-        upload_folder = app.config['UPLOAD_FOLDER']
-        applicants = [file for file in os.listdir(upload_folder) if file.endswith('.pdf')]
-        job_description = session.get('job_description', '')
-
-        # Initialize analysis_data as empty
-        analysis_data = []
-
-        return render_template("ext_recruit.html", applicants=applicants, analysis_data=analysis_data, job_description=job_description)
-
-
-
-@app.route("/ext_recruit/clear", methods=["POST"])
-@login_required
-def clear_results():
-    """Clear uploaded files and analysis results."""
-    upload_folder = app.config['UPLOAD_FOLDER']
-    db = get_db()
-    cursor = db.cursor()
-    try:
-        # Delete all files in the uploads directory
-        for filename in os.listdir(upload_folder):
-            file_path = os.path.join(upload_folder, filename)
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
-
-        # Clear database tables
-        cursor.execute("DELETE FROM applicants")
-        cursor.execute("DELETE FROM job_postings")
-        db.commit()
-
-        flash('All uploaded files and results have been cleared.')
-    except Exception as e:
-        flash(f'An error occurred while clearing files: {e}')
-    return redirect(url_for('ext_recruit'))
-
 
 @app.route("/bespoke_apology", methods=["GET", "POST"])
 @login_required
 def bespoke_apology():
-    """ Endpoint for displaying the bespoke apologies for the applicants that didn't make it to the interviews """
-    
-    # Fetch Data
+    """Display bespoke apologies for applicants not selected for interviews."""
+    # Fetch data
     db = get_db()
     cursor = db.cursor()
     cursor.execute("SELECT * FROM applicants WHERE interview_status = 'Not Selected'")
     losers = cursor.fetchall()
     cursor.execute("SELECT required_skills FROM job_postings")
     required_skills = cursor.fetchall()
-    
 
     if request.method == 'POST':
         if losers:
-            MODEL="gpt-4o-mini"
-            api_key=os.getenv("OPENAI_API_KEY", "<your OpenAI API key if not set as an env var>")
+            MODEL = "gpt-4o-mini"
+            api_key = os.getenv("OPENAI_API_KEY", "<your OpenAI API key if not set as an env var>")
 
-            # Retrieve tailored questions
+            # Retrieve bespoke apologies
             response = main.bespoke_apologies(api_key, losers, required_skills, model=MODEL)
         else:
             response = []
@@ -331,8 +158,7 @@ def bespoke_apology():
 @app.route("/tailored_interviews", methods=["GET", "POST"])
 @login_required
 def tailored_interviews():
-    """ Endpoint for displaying the tailored interviews for the applicants that made it to the interviews """
-    
+    """Display tailored interviews for selected applicants."""
     # Fetch data
     db = get_db()
     cursor = db.cursor()
@@ -341,13 +167,12 @@ def tailored_interviews():
     cursor.execute("SELECT required_skills FROM job_postings")
     required_skills = cursor.fetchall()
 
-
     if request.method == 'POST':
         if winners:
-            MODEL="gpt-4o-mini"
-            api_key=os.getenv("OPENAI_API_KEY", "<your OpenAI API key if not set as an env var>")
+            MODEL = "gpt-4o-mini"
+            api_key = os.getenv("OPENAI_API_KEY", "<your OpenAI API key if not set as an env var>")
 
-            # Retrieve tailored questions
+            # Retrieve tailored interview questions
             response = main.tailored_questions(api_key, winners, required_skills, model=MODEL)
         else:
             response = []
@@ -355,14 +180,12 @@ def tailored_interviews():
         required_skills = []
         return render_template("tailored_interviews.html", winners=winners, response=response, required_skills=required_skills)
     else:
-        return render_template("tailored_interviews.html", winners=winners, required_skills=required_skills )
-
-    
+        return render_template("tailored_interviews.html", winners=winners, required_skills=required_skills)
 
 @app.route("/uploads/<name>")
 @login_required
 def download_file(name):
-    """Send file from the uploads folder"""
+    """Send file from the uploads folder."""
     return send_from_directory(app.config['UPLOAD_FOLDER'], name)
 
 if __name__ == "__main__":
