+++ user_interface/helpers.py
@@ -1,18 +1,28 @@
-from flask import redirect, render_template, session
-from functools import wraps
+# helpers.py
 
+from flask import redirect, render_template, session, g
+from functools import wraps
+import sqlite3
+
+ALLOWED_EXTENSIONS = {'txt', 'pdf', 'png', 'jpg', 'jpeg', 'gif'}
+
+def allowed_file(filename):
+    return '.' in filename and \
+           filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
+
+def get_db():
+    if 'db' not in g:
+        g.db = sqlite3.connect('skills.db')
+        g.db.execute('PRAGMA foreign_keys = ON;')
+    return g.db
 
 def apology_login(message, code=400):
     """Render message as an apology to user."""
-
     return render_template("apology_login.html", code=code, message=message), code
-
 
 def apology_home(message, code=400):
     """Render message as an apology to user."""
-
     return render_template("apology_home.html", code=code, message=message), code
-
 
 def login_required(f):
     """
@@ -26,5 +36,3 @@ def login_required(f):
             return redirect("/login")
         return f(*args, **kwargs)
     return decorated_function
-
-
