+++ user_interface/recruitment/__init__.py
@@ -1,0 +1,3 @@
+# recruitment/__init__.py
+
+from .routes import recruitment_bp
