+++ user_interface/templates/ext_recruit.html
@@ -25,7 +25,7 @@ External Recruiting
         <!-- Upload Resumes Button -->
         <div>
             <h2>Upload Resumes</h2>
-            <form method="post" enctype="multipart/form-data">
+            <form action="{{ url_for('recruitment.upload_files') }}" method="post" enctype="multipart/form-data">
                 <input type="file" name="files" multiple required>
                 <input type="submit" value="Upload" class="action-button upload-button">
             </form>
@@ -35,7 +35,7 @@ External Recruiting
     <!-- Job Description Section (Right) -->
     <div class="right-section">
         <h2>Input Job Description</h2>
-        <form method="post" action="{{ url_for('ext_recruit') }}">
+        <form method="post" action="{{ url_for('recruitment.submit_job_description') }}">
             <textarea name="job_description" rows="10" cols="50" placeholder="Enter the job description here..." required>{{ job_description or '' }}</textarea>
             <br>
             <input type="submit" value="Save Job Description" class="action-button job-desc-button">
@@ -45,12 +45,12 @@ External Recruiting
 
 <!-- Buttons Section (Run Analysis + Clear Results) -->
 <div class="buttons-container">
-    <form method="post" action="{{ url_for('ext_recruit') }}" id="analysis-form" class="analysis-form">
+    <form method="post" action="{{ url_for('recruitment.run_analysis') }}" id="analysis-form" class="analysis-form">
         <input type="hidden" name="action" value="run_analysis">
         <button type="submit" class="action-button run-button" {% if not applicants or not job_description %}disabled{% endif %}>Run Similarity Analysis</button>
     </form>
 
-    <form method="post" action="{{ url_for('clear_results') }}" onsubmit="return confirm('Are you sure you want to delete all uploaded resumes and clear results?');">
+    <form method="post" action="{{ url_for('recruitment.clear_results') }}" onsubmit="return confirm('Are you sure you want to delete all uploaded resumes and clear results?');">
         <button type="submit" class="action-button clear-button" {% if not applicants %}disabled{% endif %}>Clear All Results</button>
     </form>
 </div>
@@ -89,13 +89,6 @@ External Recruiting
 </div>
 
 
-<!-- Clear All Results Button 
-<div style="padding: 10px 0px;">
-    <form method="post" action="{{ url_for('clear_results') }}" onsubmit="return confirm('Are you sure you want to delete all uploaded resumes and clear results?');">
-        <button type="submit" class="action-button clear-button" {% if not applicants %}disabled{% endif %}>Clear All Results</button>
-    </form>
-</div>
--->
 
 {% endblock %}
