+++ user_interface/templates/layout.html
@@ -33,7 +33,7 @@
     </div>
   </header>
   <nav class="navigation">
-    <a href="/ext_recruit" class="nav-button">
+    <a href="{{ url_for('recruitment.ext_recruit') }}" class="nav-button">
       <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-search"
         viewBox="0 0 16 16">
         <path
